Renamed ResultsDB to HDF5Storage
